Finished unit tests for create currently all passing.
